Move mappable helper methods to a new common actions module
So we can use the same customization technique overriding
the methods we need.
